value add workflows

+++ src/test/resources/db/value/value_delete_user.sql
@@ -1,5 +1,1 @@
-PREPARE value_delete_user (text, bigint, bigint) AS
-    DELETE FROM user_values
-     WHERE group_id = $1
-       AND user_id = $2
-       AND source_id = $3;
+PREPARE value_delete_user (text, bigint, bigint) AS DELETE FROM user_values WHERE group_id = $1 AND user_id = $2 AND source_id IS NOT DISTINCT FROM $3;
