+++ src/test/resources/db/value/value_delete_user_mysql.sql
@@ -1,5 +1,1 @@
-PREPARE value_delete_user FROM
-   'DELETE FROM user_values
-     WHERE group_id = ?
-       AND user_id = ?
-       AND source_id = ?';
+PREPARE value_delete_user FROM 'DELETE FROM user_values WHERE group_id = ? AND user_id = ? AND source_id <=> ?';
